using long now

+++ src/primefactors/IntegerFactorizer.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface IntegerFactorizer {
 
-	public abstract List<Integer> factorize(Integer toBeFactorized);
+	public abstract List<Long> factorize(Long toBeFactorized);
 
 }
